Refactored MarkdownConverter test suite
