+++ src/test/kotlin/com/github/woojiahao/markdown_converter/FigureElementTest.kt
@@ -9,12 +9,13 @@ class FigureElementTest {
   }
 
   @Test
+  fun `No alt text renders single img`() {
+    assertMarkdown(FIGURE_RESOURCE_NAME, "NoAltText")
+  }
+
+  @Test
   fun `Absolute web paths will generate a figure with src set to web path`() {
     assertMarkdown(FIGURE_RESOURCE_NAME, "AbsoluteWebPath")
   }
 
-  @Test
-  fun `No alt text renders single img`() {
-    assertMarkdown(FIGURE_RESOURCE_NAME, "NoAltText")
-  }
 }
