--- src/test/kotlin/com/github/woojiahao/markdown_converter/utility/HtmlUtility.kt
@@ -1,5 +1,0 @@
-package com.github.woojiahao.markdown_converter.utility
-
-fun validateDom() {
-
-}
